Breaks parser utils into it's own file.
