+++ lib/parser/utils.test.ts
@@ -1,0 +1,56 @@
+import { isClaimSame } from './utils';
+import Claim from 'lib/db/entities/claim';
+
+describe( 'isClaimSame', () => {
+	const baseClaim = {
+		number: '1234',
+		status: 'pending',
+		type: 'test',
+		slug: 'test',
+		serviceDate: new Date( 2021, 0, 1, 12 ),
+		billed: 1.23,
+		cost: 1.23,
+	} as const;
+
+	test( 'returns true on identical claims', () => {
+		const claim1 = new Claim( baseClaim );
+		const claim2 = new Claim( baseClaim );
+		expect( isClaimSame( claim1, claim2 ) ).toBeTruthy();
+	} );
+
+	function getDiffOnKey(
+		claimKey: keyof typeof baseClaim
+	): string | Date | number {
+		switch ( typeof baseClaim[ claimKey ] ) {
+			case 'string':
+				return 'changed';
+			case 'object':
+				return new Date( 2000, 1 );
+			case 'number':
+				return 4.56;
+			default:
+				throw new Error(
+					'Unknown type: ' + typeof baseClaim[ claimKey ]
+				);
+		}
+	}
+
+	test( 'returns false on claim with different keys', () => {
+		Object.keys( baseClaim )
+			.filter( ( claimKey ) => ! [ 'slug' ].includes( claimKey ) )
+			.forEach( ( claimKey: keyof typeof baseClaim ) => {
+				const claim1 = new Claim( { ...baseClaim } );
+				const claim2 = new Claim( {
+					...baseClaim,
+					[ claimKey ]: getDiffOnKey( claimKey ),
+				} );
+				expect( isClaimSame( claim1, claim2 ) ).toBeFalsy();
+			} );
+	} );
+} );
+
+describe( 'getHash', () => {
+	test.todo( 'gets correct hash for empty array' );
+	test.todo( 'gets correct hash with dummy data' );
+	test.todo( 'returns length from second arg' );
+} );
